+++ tui/src/tui/md_parser/types.rs
@@ -34,7 +34,7 @@ pub type FragmentsInOneLine<'a> = MdLineFragments<'a>;
 /// Alias for [List] of [FragmentsInOneLine].
 pub type Lines<'a> = List<FragmentsInOneLine<'a>>;
 
-#[derive(Clone, Debug, PartialEq, size_of::SizeOf)]
+#[derive(Clone, Debug, PartialEq)]
 pub struct HeadingData<'a> {
     pub heading_level: HeadingLevel,
     pub text: &'a str,
@@ -48,7 +48,7 @@ pub struct HeadingData<'a> {
 /// - There are some exceptions such as smart lists and code blocks which represent
 ///   multiple lines of text.
 #[allow(clippy::large_enum_variant)]
-#[derive(Clone, Debug, PartialEq, size_of::SizeOf)]
+#[derive(Clone, Debug, PartialEq)]
 pub enum MdBlock<'a> {
     Heading(HeadingData<'a>),
     SmartList((Lines<'a>, BulletKind, usize)),
@@ -62,7 +62,7 @@ pub enum MdBlock<'a> {
 
 /// These are things that show up in a single line of Markdown text [MdLineFragments]. They do not
 /// include other Markdown blocks (like code blocks, lists, headings, etc).
-#[derive(Clone, Debug, PartialEq, size_of::SizeOf)]
+#[derive(Clone, Debug, PartialEq)]
 pub enum MdLineFragment<'a> {
     UnorderedListBullet {
         indent: usize,
@@ -82,7 +82,7 @@ pub enum MdLineFragment<'a> {
     Checkbox(bool),
 }
 
-#[derive(Clone, Debug, PartialEq, size_of::SizeOf)]
+#[derive(Clone, Debug, PartialEq)]
 pub struct HyperlinkData<'a> {
     pub text: &'a str,
     pub url: &'a str,
@@ -100,7 +100,7 @@ mod hyperlink_data_impl {
     }
 }
 
-#[derive(Copy, Clone, Debug, PartialEq, Eq, size_of::SizeOf)]
+#[derive(Copy, Clone, Debug, PartialEq, Eq)]
 pub struct HeadingLevel {
     pub level: usize,
 }
@@ -160,7 +160,7 @@ pub mod constants {
     pub const EXCLAMATION: &str = "!";
 }
 
-#[derive(Debug, PartialEq, Clone, size_of::SizeOf)]
+#[derive(Debug, PartialEq, Clone)]
 pub struct CodeBlockLine<'a> {
     pub language: Option<&'a str>,
     pub content: CodeBlockLineContent<'a>,
@@ -169,7 +169,7 @@ pub struct CodeBlockLine<'a> {
 /// Alias for [List] of [CodeBlockLine].
 pub type CodeBlockLines<'a> = List<CodeBlockLine<'a>>;
 
-#[derive(Debug, PartialEq, Clone, size_of::SizeOf)]
+#[derive(Debug, PartialEq, Clone)]
 pub enum CodeBlockLineContent<'a> {
     Text(&'a str),
     StartTag,